+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './book';
 export * from './number';
 export * from './record';
+export * from './validation';
